Fix some compiler warnings.

+++ src/input.c
@@ -65,7 +65,7 @@ static OSTime lastButtonPress = 0;
 
 typedef struct
 {
-    int globalMaxlength;
+    size_t globalMaxlength;
     bool globalLimit;
     bool okButtonEnabled;
     OSThread *calcThread;
@@ -129,7 +129,7 @@ static void SWKBD_Render(SWKBD_Args *args, KeyboardChecks check)
                 }
         }
 
-        args->okButtonEnabled = args->globalLimit ? len == args->globalMaxlength : len <= args->globalMaxlength; // This rises a harmless compiler warning
+        args->okButtonEnabled = args->globalLimit ? len == args->globalMaxlength : len <= args->globalMaxlength;
     }
     else
         args->okButtonEnabled = false;
